refactor: extract Express server setup into module
